requirements.txt
google-cloud-bigquery==3.27.0
google-cloud-storage==2.19.0
google-cloud-vision==3.8.1
google-genai==1.9.0
requests==2.32.3
beautifulsoup4==4.12.3
pandas==2.2.3
pyarrow==18.1.0
python-dotenv==1.0.1
db-dtypes==1.3.1
